pyproject.toml
[tool.poetry]
name = "chaosengine-bot"
version = "0.2.0"
description = "A discord bot for Diablo:Immortal clans and warbands"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "chaosengine_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.22.2"
"discord.py" = "^2.1.0"
colorama = "^0.4.6"
discord = "^2.1.0"
mysql-connector-python = "^8.0.31"
python-dotenv = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"